Adds cold boot restart to load test proj
