+++ src/Umbraco.TestData/Umbraco.TestData.csproj
@@ -41,6 +41,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="LoadTestComponent.cs" />
+    <Compile Include="LoadTestComposer.cs" />
     <Compile Include="LoadTestController.cs" />
     <Compile Include="SegmentTestController.cs" />
     <Compile Include="UmbracoTestDataController.cs" />
